Modal component in a separate file
