+++ src/components/EventList/eventlist.scss
@@ -69,111 +69,49 @@
   }
 }
 
-.popupcontainer{
-  position: fixed;
-  top: 0em;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  overflow-y: scroll;
+.modalimage {
+  img{
+    width: 100%;
+  }
 }
-
-.popup {
-  position: relative;
-  z-index: 400;
-  margin: auto;
-  top: 0;
-  width: 50%;
-  border-radius: 5px;
-  background: rgba(255, 255, 255, 0.9);
-  margin-bottom: 3em;
-  margin-top: 3em;
+.modal-event-property{
+  padding-top: 1em;
   padding-bottom: 1em;
-}
-
-.shade {
-  position: fixed;
-  z-index: 350;
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  background: rgba(0, 0, 0, 0.8);
-  .shadecontent{
-    position: absolute;
-    top: 1em;
-    right: 2em;
-    .cross{
-      color: $text-color-light;
-    }
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  .icon{
+    width: 25px;
+    height: 25px;
   }
-}
-
-.modalcontent {
-  overflow-wrap: break-word;
-  text-align: justify;
-
-  .modalimage {
-    img{
-      width: 100%;
-      }
-  }
-  .modal-event-property{
-      padding-top: 1em;
-      padding-bottom: 1em;
-      display: flex;
-      flex-direction: row;
-      justify-content: space-around;
-      .icon{
-          width: 25px;
-          height: 25px;
-      }
-      .icon_group{
-        text-align: center;
-        align-items: center;
-        display: flex;
-        flex-direction: column;
-      }
-      p{
-          margin: 0 0 0 5px;
-      }
-  }
-  .modalinfo{
-    padding: 1em 1em;
-  }
-  .modalbutton{
+  .icon_group{
     text-align: center;
-    .rsvpbutton{
-      display: inline-block;
-        border-radius: 4px;
-        background-color: $melon;
-        border: none;
-        color: #FFFFFF;
-        text-align: center;
-        font-size: 16px;
-        padding: 20px;
-        transition: all 0.5s;
-        margin: 5px;
-      }
-      .rsvpbutton:hover{
-        text-decoration: underline;
-      }
-      .rsvpclosed{
-        display: inline-block;
-          border-radius: 4px;
-          background-color: #f4511e;
-          border: none;
-          color: #FFFFFF;
-          text-align: center;
-          font-size: 16px;
-          padding: 20px;
-          transition: all 0.5s;
-          margin: 5px;
-        }
-    }
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+  }
+  p{
+    margin: 0 0 0 5px;
+  }
 }
+.modalinfo{
+  padding: 1em 1em;
+}
+
+
 
 @media screen and (max-width: $breakpoint-mobile) {
+
+  .description{
+    font-size: 16px;
+  }
+
+  .modal-event-property{
+    flex-direction: column;
+    .icon_group{
+      flex-direction: row;
+    }
+  }
 
   .events-feed {
     padding: 0%;
@@ -197,40 +135,18 @@
   .name {
     font-size: large;
   }
-  .popup {
-    width: 75%;
-  }
-  .modalcontent {
-    .description{
-      font-size: 16px;
-    }
-    .modal-event-property{
-        flex-direction: column;
-        .icon_group{
-          flex-direction: row;
-        }
-      }
-    }
-    .shade {
-      .shadecontent{
-        position: absolute;
-        top: 0px;
-        right: 5px;
-        .cross{
-          color: $text-color-light;
-        }
-      }
-    }
+
 }
 
 // flipped phone
 @media screen and (max-width: $breakpoint-mobile-flipped) {
-  .modalcontent {
     .modal-event-property{
-        flex-direction: column;
-        .icon_group{
-          flex-direction: row;
-        }
+      flex-direction: column;
+      .icon_group{
+        flex-direction: row;
       }
     }
   }
+
+
+
