+++ src/layouts/Exhibitors/index.js
@@ -1,17 +1,20 @@
 import React, {PropTypes} from "react"
 
 import ExhibitorList from "../../components/ExhibitorList"
+import "./index.scss"
 
 
 const Exhibitors = (props) => {
 
   return (
 
-    <div>
+   <div className="content">
+    <div className= "Exhibitors-container">
 
     <ExhibitorList {...props}/>
 
      </div>
+   </div>
   )
 }
